# docs/specs/api_contract.yaml
openapi: 3.1.0
info:
  title: Derived API Contract
  version: 1.0.0
paths:
...
  /equipment:
    delete:
      operationId: delete_equipment
      responses:
        '204':
          description: Derived from SPEC.md
    get:
      operationId: get_equipment
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: post_equipment
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /equipment/:id:
    delete:
      operationId: delete_equipment_id
      responses:
        '204':
          description: Derived from SPEC.md
    get:
      operationId: get_equipment_id
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
    patch:
      operationId: patch_equipment_id
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /loans:
    get:
      operationId: get_loans
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: post_loans
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /loans/:id/approve:
    post:
      operationId: post_loans_id_approve
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /loans/:id/cancel:
    post:
      operationId: post_loans_id_cancel
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /loans/:id/reject:
    post:
      operationId: post_loans_id_reject
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /loans/:id/return:
    post:
      operationId: post_loans_id_return
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...